+++ rust-src/concordium_base/src/ffi_helpers/common.rs
@@ -6,3 +6,11 @@ extern "C" fn free_array_len(ptr: *mut u8, len: u64) {
         Vec::from_raw_parts(ptr, len as usize, len as usize);
     }
 }
+
+#[no_mangle]
+/// Free a vector using its raw pointer, length and capacity.
+extern "C" fn free_array_len_cap(ptr: *mut u8, len: u64, cap: u64) {
+    unsafe {
+        Vec::from_raw_parts(ptr, len as usize, cap as usize);
+    }
+}